+++ nistbeacon/nistbeacon.py
@@ -1,17 +1,17 @@
 import requests
 from requests.exceptions import RequestException
 
-import py_nist_beacon.nist_beacon_constants as cn
-from py_nist_beacon.nist_randomness_beacon_value import (
-    NistRandomnessBeaconValue
+import nistbeacon.constants as cn
+from nistbeacon.nistbeaconvalue import (
+    NistBeaconValue
 )
 
 
-class NistRandomnessBeacon(object):
+class NistBeacon(object):
     NIST_BASE_URL = "https://beacon.nist.gov/rest/record"
 
     @classmethod
-    def _query_nist(cls, url_data: str) -> NistRandomnessBeaconValue:
+    def _query_nist(cls, url_data: str) -> NistBeaconValue:
         try:
             r = requests.get(
                 "{0}/{1}".format(
@@ -21,7 +21,7 @@ class NistRandomnessBeacon(object):
             )
 
             if r.status_code is requests.codes.OK:
-                return NistRandomnessBeaconValue.from_xml(r.text)
+                return NistBeaconValue.from_xml(r.text)
             else:
                 return None
         except RequestException:
@@ -42,7 +42,7 @@ class NistRandomnessBeacon(object):
 
         record = cls.get_record(timestamp)
 
-        if isinstance(record, NistRandomnessBeaconValue) is False:
+        if isinstance(record, NistBeaconValue) is False:
             # Don't you dare try to play me
             return False
 
@@ -55,14 +55,14 @@ class NistRandomnessBeacon(object):
             return False
 
         if (
-            isinstance(prev_record, NistRandomnessBeaconValue) and
-            isinstance(next_record, NistRandomnessBeaconValue)
+            isinstance(prev_record, NistBeaconValue) and
+            isinstance(next_record, NistBeaconValue)
         ):
             # Majority case, somewhere in the middle of the chain
             # True if:
             #   - All three records have proper signatures
             #   - The requested record's previous output equals previous
-            #   - The next possible record's previosu output equals the record
+            #   - The next possible record's previous output equals the record
             return (
                 record.valid_signature and
                 prev_record.valid_signature and
@@ -73,10 +73,10 @@ class NistRandomnessBeacon(object):
 
         if (
             prev_record is None and
-            isinstance(next_record, NistRandomnessBeaconValue)
+            isinstance(next_record, NistBeaconValue)
         ):
             # Edge case, this was potentially the first record of all time
-            init_ref = NistRandomnessBeaconValue.from_json(
+            init_ref = NistBeaconValue.from_json(
                 cn.NIST_INIT_RECORD,
             )
 
@@ -88,7 +88,7 @@ class NistRandomnessBeacon(object):
             )
 
         if (
-            isinstance(prev_record, NistRandomnessBeaconValue) and
+            isinstance(prev_record, NistBeaconValue) and
             next_record is None
         ):
             # Edge case, this was potentially the latest and greatest
@@ -102,7 +102,7 @@ class NistRandomnessBeacon(object):
     def get_first_record(
             cls,
             download: bool=False
-    ) -> NistRandomnessBeaconValue:
+    ) -> NistBeaconValue:
         """
         Get the first (oldest) record available. Since the first record
         IS a known value in the system we can load it from constants.
@@ -113,14 +113,14 @@ class NistRandomnessBeacon(object):
         """
 
         if download:
-            return NistRandomnessBeacon.get_record(
+            return NistBeacon.get_record(
                 cn.NIST_INIT_RECORD_TIMESTAMP
             )
         else:
-            return NistRandomnessBeaconValue.from_json(cn.NIST_INIT_RECORD)
+            return NistBeaconValue.from_json(cn.NIST_INIT_RECORD)
 
     @classmethod
-    def get_last_record(cls) -> NistRandomnessBeaconValue:
+    def get_last_record(cls) -> NistBeaconValue:
         """
         Get the last (newest) record available.
 
@@ -130,7 +130,7 @@ class NistRandomnessBeacon(object):
         return cls._query_nist("last")
 
     @classmethod
-    def get_next(cls, timestamp: int) -> NistRandomnessBeaconValue:
+    def get_next(cls, timestamp: int) -> NistBeaconValue:
         """
         Given a record timestamp, get the next record available.
 
@@ -141,7 +141,7 @@ class NistRandomnessBeacon(object):
         return cls._query_nist("next/{}".format(timestamp))
 
     @classmethod
-    def get_previous(cls, timestamp: int) -> NistRandomnessBeaconValue:
+    def get_previous(cls, timestamp: int) -> NistBeaconValue:
         """
         Given a record timestamp, get the previous record available.
 
@@ -152,7 +152,7 @@ class NistRandomnessBeacon(object):
         return cls._query_nist("previous/{}".format(timestamp))
 
     @classmethod
-    def get_record(cls, timestamp: int) -> NistRandomnessBeaconValue:
+    def get_record(cls, timestamp: int) -> NistBeaconValue:
         """
         Get a specific record (or next closest)