+++ py_nist_beacon/__init__.py
@@ -1,3 +1,0 @@
-from py_nist_beacon.nist_randomness_beacon import NistRandomnessBeacon
-
-__version__ = '0.6.0'